requirements.txt
fastapi==0.115.12
SQLAlchemy==2.0.41
dependency-injector==4.46.0
pytest==8.3.5
uvicorn==0.34.2
httpx==0.28.1
Faker==37.3.0
alembic==1.15.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
